tests: split test_config.py into focused modules (P3-F05, F06, F08, F17)
- Remove test_loaded_config_has_sonnet_default (strict subset of test_default_model_config)
- Remove test_fleet_config_default_l2_timeout and test_defaults_yaml_has_fleet_section (subsumed by test_fleet_l2_timeout_matches_defaults_yaml)
- Replace `assert FleetConfig is not None` tautology with `assert dataclasses.is_dataclass(FleetConfig)`
- Extract TestFleetConfig, TestSkillsConfig, TestSubsetsConfig, TestPacksConfig, TestWriteConfigLayer into dedicated files; test_config.py shrinks from 1317 to 856 lines
