+++ tests/config/test_config.py
@@ -1,7 +1,6 @@
 """Tests for configuration loading and resolution."""
 
 from dataclasses import fields as dc_fields
-from pathlib import Path
 
 import pytest
 import yaml
@@ -201,11 +200,6 @@ class TestLoadConfig:
         (config_dir / "config.yaml").write_text(yaml.dump(config_data))
         cfg = load_config(tmp_path)
         assert cfg.model.override == "haiku"
-        assert cfg.model.default == "sonnet"
-
-    def test_loaded_config_has_sonnet_default(self, tmp_path):
-        """MOD_C4: load_config produces model.default='sonnet'."""
-        cfg = load_config(tmp_path)
         assert cfg.model.default == "sonnet"
 
     def test_yaml_loads_worktree_setup_config(self, tmp_path):
@@ -840,303 +834,6 @@ class TestBranchingConfig:
         assert cfg.branching.promotion_target == "stable"
 
 
-class TestBuildSubsetsConfigCustomTagsValidation:
-    """CC-F2: _build_subsets_config must raise ValueError for non-dict custom_tags."""
-
-    @pytest.mark.parametrize("bad_value", [["list", "not", "dict"], "oops", 42])
-    def test_build_subsets_config_raises_for_non_dict_custom_tags(self, bad_value: object) -> None:
-        """Non-dict custom_tags must raise ValueError, not silently coerce to {}."""
-        from autoskillit.config.settings import _build_subsets_config
-
-        with pytest.raises(ValueError, match="custom_tags"):
-            _build_subsets_config({"custom_tags": bad_value})
-
-    def test_build_subsets_config_dict_custom_tags_accepted(self):
-        """Valid dict custom_tags must not raise."""
-        from autoskillit.config.settings import _build_subsets_config
-
-        result = _build_subsets_config({"custom_tags": {"my_tag": ["skill-a"]}})
-        assert result.custom_tags == {"my_tag": ["skill-a"]}
-
-    def test_build_subsets_config_empty_dict_custom_tags_accepted(self):
-        """Empty dict custom_tags must not raise."""
-        from autoskillit.config.settings import _build_subsets_config
-
-        result = _build_subsets_config({"custom_tags": {}})
-        assert result.custom_tags == {}
-
-
-class TestSkillsConfig:
-    """SkillsConfig dataclass, tier duplication validation, and AutomationConfig integration."""
-
-    def test_skills_config_dataclass(self) -> None:
-        """SkillsConfig has tier1, tier2, tier3 list[str] fields."""
-        from autoskillit.config.settings import SkillsConfig
-
-        sc = SkillsConfig(tier1=["a"], tier2=["b"], tier3=["c"])
-        assert sc.tier1 == ["a"] and sc.tier2 == ["b"] and sc.tier3 == ["c"]
-
-    def test_skills_config_tier_duplication_raises(self) -> None:
-        """Skill in multiple tiers raises ValueError at construction (REQ-TIER-009)."""
-        import pytest
-
-        from autoskillit.config.settings import SkillsConfig
-
-        with pytest.raises(ValueError, match="multiple tiers"):
-            SkillsConfig(tier1=["open-kitchen"], tier2=["open-kitchen"], tier3=[])
-
-    def test_automation_config_has_skills_field(self) -> None:
-        """AutomationConfig.skills is a SkillsConfig with tier lists."""
-        from autoskillit.config.settings import AutomationConfig, SkillsConfig
-
-        cfg = AutomationConfig()
-        assert isinstance(cfg.skills, SkillsConfig)
-
-    def test_defaults_yaml_skills_section(self) -> None:
-        """defaults.yaml has non-empty skills.tier1/tier2/tier3 lists."""
-        from autoskillit.core import load_yaml, pkg_root
-
-        defaults = load_yaml(pkg_root() / "config" / "defaults.yaml")
-        assert isinstance(defaults, dict)
-        skills = defaults.get("skills", {})
-        assert "open-kitchen" in skills.get("tier1", [])
-        assert len(skills.get("tier2", [])) >= 20
-        assert len(skills.get("tier3", [])) >= 10
-
-    def test_load_config_populates_skills_tiers(self, tmp_path) -> None:
-        """load_config() produces an AutomationConfig with tier assignments from defaults."""
-        from autoskillit.config import load_config
-
-        cfg = load_config(tmp_path / "settings.toml")
-        assert "open-kitchen" in cfg.skills.tier1
-        assert "make-plan" in cfg.skills.tier2
-        assert "compose-pr" in cfg.skills.tier3
-
-    def test_skills_config_exported_from_config_package(self) -> None:
-        """SkillsConfig is importable from autoskillit.config and has expected fields."""
-        from autoskillit.config import SkillsConfig
-
-        cfg = SkillsConfig()
-        assert hasattr(cfg, "tier1")
-        assert hasattr(cfg, "tier2")
-        assert hasattr(cfg, "tier3")
-
-
-class TestSubsetsConfig:
-    # T1 — SubsetsConfig defaults
-
-    def test_subsets_config_default_disabled_is_empty_list(self) -> None:
-        from autoskillit.config import SubsetsConfig
-
-        cfg = SubsetsConfig()
-        assert cfg.disabled == []
-
-    def test_subsets_config_default_custom_tags_is_empty_dict(self) -> None:
-        from autoskillit.config import SubsetsConfig
-
-        cfg = SubsetsConfig()
-        assert cfg.custom_tags == {}
-
-    def test_automation_config_has_subsets_field(self) -> None:
-        from autoskillit.config import AutomationConfig, SubsetsConfig
-
-        cfg = AutomationConfig()
-        assert isinstance(cfg.subsets, SubsetsConfig)
-        assert cfg.subsets.disabled == []
-        assert cfg.subsets.custom_tags == {}
-
-    # T2 — load_config with subsets.disabled
-
-    def test_load_config_subsets_disabled(self, tmp_path) -> None:
-        config_dir = tmp_path / ".autoskillit"
-        config_dir.mkdir()
-        (config_dir / "config.yaml").write_text("subsets:\n  disabled:\n    - github\n    - ci\n")
-        cfg = load_config(tmp_path)
-        assert cfg.subsets.disabled == ["github", "ci"]
-
-    def test_load_config_subsets_disabled_absent_means_empty(self, tmp_path) -> None:
-        cfg = load_config(tmp_path)
-        assert cfg.subsets.disabled == []
-
-    # T3 — load_config with subsets.custom_tags
-
-    def test_load_config_subsets_custom_tags(self, tmp_path) -> None:
-        config_dir = tmp_path / ".autoskillit"
-        config_dir.mkdir()
-        yaml_text = (
-            "subsets:\n"
-            "  custom_tags:\n"
-            "    my-team-tools:\n"
-            "      - investigate\n"
-            "      - make-plan\n"
-        )
-        (config_dir / "config.yaml").write_text(yaml_text)
-        cfg = load_config(tmp_path)
-        assert cfg.subsets.custom_tags == {"my-team-tools": ["investigate", "make-plan"]}
-
-    def test_load_config_subsets_custom_tags_absent_means_empty(self, tmp_path) -> None:
-        cfg = load_config(tmp_path)
-        assert cfg.subsets.custom_tags == {}
-
-    # T4 — Unknown category warning, no crash
-
-    def test_load_config_unknown_disabled_category_logs_warning_not_crash(self, tmp_path) -> None:
-        import logging
-
-        config_dir = tmp_path / ".autoskillit"
-        config_dir.mkdir()
-        (config_dir / "config.yaml").write_text(
-            "subsets:\n  disabled:\n    - totally-unknown-category\n"
-        )
-        # Attach a handler directly to the logger to capture warnings reliably.
-        # caplog is unreliable here because the structlog capture_logs() autouse
-        # fixture can intercept the handler chain under xdist worker ordering.
-        captured: list[logging.LogRecord] = []
-        handler = logging.Handler()
-        handler.emit = captured.append  # type: ignore[assignment]
-        logger = logging.getLogger("autoskillit.config.settings")  # noqa: TID251
-        logger.addHandler(handler)
-        try:
-            cfg = load_config(tmp_path)
-        finally:
-            logger.removeHandler(handler)
-        assert cfg.subsets.disabled == ["totally-unknown-category"]  # preserved as-is
-        assert any("totally-unknown-category" in r.getMessage() for r in captured)
-
-    def test_load_config_known_disabled_category_no_warning(self, tmp_path, caplog) -> None:
-        import logging
-
-        config_dir = tmp_path / ".autoskillit"
-        config_dir.mkdir()
-        (config_dir / "config.yaml").write_text("subsets:\n  disabled:\n    - github\n")
-        with caplog.at_level(logging.WARNING):
-            load_config(tmp_path)
-        assert not any("github" in r.message for r in caplog.records)
-
-    def test_load_config_custom_tag_in_disabled_is_valid(self, tmp_path, caplog) -> None:
-        """Custom tags defined in custom_tags can also appear in disabled."""
-        import logging
-
-        config_dir = tmp_path / ".autoskillit"
-        config_dir.mkdir()
-        yaml_text = (
-            "subsets:\n"
-            "  disabled:\n"
-            "    - experimental\n"
-            "  custom_tags:\n"
-            "    experimental:\n"
-            "      - write-recipe\n"
-        )
-        (config_dir / "config.yaml").write_text(yaml_text)
-        with caplog.at_level(logging.WARNING):
-            load_config(tmp_path)
-        assert not any("experimental" in r.message for r in caplog.records)
-
-    # T5 — SubsetsConfig exported from config package
-
-    def test_subsets_config_importable_from_config_package(self) -> None:
-        from autoskillit.config import SubsetsConfig
-
-        cfg = SubsetsConfig()
-        assert hasattr(cfg, "disabled")
-        assert hasattr(cfg, "custom_tags")
-
-
-class TestPacksConfig:
-    # REQ-PACK-003: PacksConfig.enabled default
-
-    def test_packs_config_default_enabled_is_empty_list(self) -> None:
-        from autoskillit.config import PacksConfig
-
-        assert PacksConfig().enabled == []
-
-    def test_automation_config_has_packs_field(self) -> None:
-        from autoskillit.config import AutomationConfig, PacksConfig
-
-        cfg = AutomationConfig()
-        assert isinstance(cfg.packs, PacksConfig)
-        assert cfg.packs.enabled == []
-
-    def test_load_config_packs_enabled(self, tmp_path) -> None:
-        from autoskillit.config import load_config
-
-        config_dir = tmp_path / ".autoskillit"
-        config_dir.mkdir()
-        (config_dir / "config.yaml").write_text("packs:\n  enabled:\n    - research\n")
-        config = load_config(tmp_path)
-        assert config.packs.enabled == ["research"]
-
-    def test_load_config_packs_enabled_absent_means_empty(self, tmp_path) -> None:
-        from autoskillit.config import load_config
-
-        config_dir = tmp_path / ".autoskillit"
-        config_dir.mkdir()
-        (config_dir / "config.yaml").write_text("subsets:\n  disabled: []\n")
-        config = load_config(tmp_path)
-        assert config.packs.enabled == []
-
-    def test_load_config_unknown_pack_in_packs_enabled_logs_warning(self, tmp_path) -> None:
-        import logging
-
-        from autoskillit.config import load_config
-
-        config_dir = tmp_path / ".autoskillit"
-        config_dir.mkdir()
-        (config_dir / "config.yaml").write_text("packs:\n  enabled:\n    - nonexistent-pack\n")
-        captured: list[logging.LogRecord] = []
-        handler = logging.Handler()
-        handler.emit = captured.append  # type: ignore[assignment]
-        logger = logging.getLogger("autoskillit.config.settings")  # noqa: TID251
-        logger.addHandler(handler)
-        try:
-            config = load_config(tmp_path)
-        finally:
-            logger.removeHandler(handler)
-        assert config.packs.enabled == ["nonexistent-pack"]  # preserved as-is
-        assert any("nonexistent-pack" in r.getMessage() for r in captured)
-
-    def test_write_config_layer_accepts_packs_enabled(self, tmp_path) -> None:
-        """write_config_layer does not raise for valid packs.enabled."""
-        from autoskillit.config.settings import write_config_layer
-
-        config_path = tmp_path / "config.yaml"
-        write_config_layer(config_path, {"packs": {"enabled": ["research"]}})
-        assert config_path.exists()
-
-
-class TestWriteConfigLayer:
-    def test_write_config_layer_rejects_secret_key(self, tmp_path: Path) -> None:
-        """write_config_layer raises ConfigSchemaError before touching the file."""
-
-        from autoskillit.config.settings import ConfigSchemaError, write_config_layer
-
-        config_path = tmp_path / "config.yaml"
-        with pytest.raises(ConfigSchemaError, match="github.token"):
-            write_config_layer(config_path, {"github": {"token": "ghp_test"}})
-        assert not config_path.exists(), "config.yaml must not be written on schema error"
-
-    def test_write_config_layer_rejects_unknown_key(self, tmp_path: Path) -> None:
-        """write_config_layer raises ConfigSchemaError for unknown section."""
-        from autoskillit.config.settings import ConfigSchemaError, write_config_layer
-
-        config_path = tmp_path / "config.yaml"
-        with pytest.raises(ConfigSchemaError, match="unrecognized key"):
-            write_config_layer(config_path, {"invented_section": {"foo": "bar"}})
-        assert not config_path.exists()
-
-    def test_write_config_layer_writes_valid_content(self, tmp_path: Path) -> None:
-        """write_config_layer writes valid schema content atomically."""
-        import yaml as _yaml
-
-        from autoskillit.config.settings import write_config_layer
-
-        config_path = tmp_path / "config.yaml"
-        write_config_layer(config_path, {"github": {"default_repo": "owner/repo"}})
-        assert config_path.is_file()
-        data = _yaml.safe_load(config_path.read_text())
-        assert data["github"]["default_repo"] == "owner/repo"
-
-
 class TestWorkspaceConfig:
     """WorkspaceConfig section is present in AutomationConfig with correct defaults."""
 
@@ -1157,161 +854,3 @@ class TestWorkspaceConfig:
 
         cfg = load_config(tmp_path / "settings.toml")
         assert cfg.workspace.runs_root is None
-
-
-class TestFleetConfig:
-    def test_fleet_config_default_l2_timeout(self) -> None:
-        """FleetConfig.default_timeout_sec defaults to 3600."""
-        from autoskillit.config.settings import FleetConfig
-
-        cfg = FleetConfig()
-        assert cfg.default_timeout_sec == 3600
-
-    def test_automation_config_has_fleet_field(self) -> None:
-        """AutomationConfig exposes fleet as a FleetConfig."""
-        from autoskillit.config.settings import AutomationConfig, FleetConfig
-
-        cfg = AutomationConfig()
-        assert isinstance(cfg.fleet, FleetConfig)
-        assert cfg.fleet.default_timeout_sec == 3600
-
-    def test_defaults_yaml_has_fleet_section(self) -> None:
-        """defaults.yaml defines fleet.default_timeout_sec."""
-        from autoskillit.core.io import load_yaml
-        from autoskillit.core.paths import pkg_root
-
-        defaults = load_yaml(pkg_root() / "config" / "defaults.yaml")
-        assert "fleet" in defaults
-        assert defaults["fleet"]["default_timeout_sec"] == 3600
-
-    def test_fleet_l2_timeout_matches_defaults_yaml(self) -> None:
-        """FleetConfig Python default matches defaults.yaml value."""
-        from autoskillit.config.settings import FleetConfig
-        from autoskillit.core.io import load_yaml
-        from autoskillit.core.paths import pkg_root
-
-        defaults = load_yaml(pkg_root() / "config" / "defaults.yaml")
-        yaml_val = defaults["fleet"]["default_timeout_sec"]
-        assert FleetConfig().default_timeout_sec == yaml_val
-
-    def test_load_config_fleet_override(self, tmp_path) -> None:
-        """User config with fleet section overrides the default."""
-        import yaml
-
-        from autoskillit.config import load_config
-
-        config_dir = tmp_path / ".autoskillit"
-        config_dir.mkdir()
-        config_data = {"fleet": {"default_timeout_sec": 7200}}
-        (config_dir / "config.yaml").write_text(yaml.dump(config_data))
-        cfg = load_config(tmp_path)
-        assert cfg.fleet.default_timeout_sec == 7200
-
-    def test_fleet_config_importable_from_config_package(self) -> None:
-        """FleetConfig is importable from autoskillit.config."""
-        from autoskillit.config import FleetConfig
-
-        assert FleetConfig is not None
-
-    def test_fleet_key_accepted_by_schema_validator(self, tmp_path) -> None:
-        """User config with fleet: section does not raise ConfigSchemaError."""
-        import yaml
-
-        from autoskillit.config import load_config
-
-        config_dir = tmp_path / ".autoskillit"
-        config_dir.mkdir()
-        config_data = {"fleet": {"default_timeout_sec": 1800}}
-        (config_dir / "config.yaml").write_text(yaml.dump(config_data))
-        cfg = load_config(tmp_path)  # must not raise ConfigSchemaError
-        assert cfg.fleet.default_timeout_sec == 1800
-
-    def test_fleet_config_rejects_zero_timeout(self) -> None:
-        """FleetConfig raises ValueError when default_timeout_sec is zero."""
-        import pytest
-
-        from autoskillit.config.settings import FleetConfig
-
-        with pytest.raises(ValueError, match="default_timeout_sec must be positive"):
-            FleetConfig(default_timeout_sec=0).validate(True)
-
-    def test_fleet_config_rejects_negative_timeout(self) -> None:
-        """FleetConfig raises ValueError when default_timeout_sec is negative."""
-        import pytest
-
-        from autoskillit.config.settings import FleetConfig
-
-        with pytest.raises(ValueError, match="default_timeout_sec must be positive"):
-            FleetConfig(default_timeout_sec=-1).validate(True)
-
-    def test_fleet_config_validate_skips_when_feature_disabled(self) -> None:
-        """FC_NEW_2: validate(False) does NOT raise even for invalid timeout."""
-        from autoskillit.config.settings import FleetConfig
-
-        FleetConfig(default_timeout_sec=0).validate(False)  # must not raise
-
-    def test_fleet_config_construction_no_longer_raises_for_invalid_timeout(self) -> None:
-        """FC_NEW_3: FleetConfig(default_timeout_sec=0) constructs without raising."""
-        from autoskillit.config.settings import FleetConfig
-
-        cfg = FleetConfig(default_timeout_sec=0)
-        assert cfg.default_timeout_sec == 0
-
-    def test_load_config_fleet_invalid_timeout_skips_when_disabled(self, tmp_path: Path) -> None:
-        """FC_NEW_4: load_config does NOT raise with invalid timeout when fleet disabled."""
-        config_dir = tmp_path / ".autoskillit"
-        config_dir.mkdir()
-        config_data = {
-            "fleet": {"default_timeout_sec": -1},
-            "features": {"fleet": False},
-        }
-        (config_dir / "config.yaml").write_text(yaml.dump(config_data))
-        cfg = load_config(tmp_path)  # must not raise
-        assert cfg.fleet.default_timeout_sec == -1
-
-    def test_load_config_fleet_invalid_timeout_raises_when_enabled(self, tmp_path: Path) -> None:
-        """FC_NEW_5: load_config raises ValueError with invalid timeout when fleet enabled."""
-        config_dir = tmp_path / ".autoskillit"
-        config_dir.mkdir()
-        config_data = {
-            "fleet": {"default_timeout_sec": -1},
-            "features": {"fleet": True},
-        }
-        (config_dir / "config.yaml").write_text(yaml.dump(config_data))
-        with pytest.raises(ValueError, match="default_timeout_sec must be positive"):
-            load_config(tmp_path)
-
-
-def test_project_config_experimental_enabled_or_fleet_enabled() -> None:
-    """Integration's resolved config enables fleet via experimental_enabled blanket."""
-    from pathlib import Path
-
-    from autoskillit.config.settings import load_config
-    from autoskillit.core.feature_flags import is_feature_enabled
-
-    project_root = Path(__file__).resolve().parents[2]
-    if not (project_root / ".autoskillit" / "config.yaml").exists():
-        pytest.skip("project config not present (clean clone or CI)")
-    cfg = load_config(project_root)
-    assert (
-        is_feature_enabled("fleet", cfg.features, experimental_enabled=cfg.experimental_enabled)
-        is True
-    )
-
-
-def test_config_resolution_fleet_enabled_via_experimental() -> None:
-    """Full config resolution enables fleet via experimental_enabled=True from defaults."""
-    from pathlib import Path
-
-    from autoskillit.config.settings import load_config
-    from autoskillit.core.feature_flags import is_feature_enabled
-
-    project_root = Path(__file__).resolve().parents[2]
-    if not (project_root / ".autoskillit" / "config.yaml").exists():
-        pytest.skip("project config not present (clean clone or CI)")
-    cfg = load_config(project_root)
-    assert cfg.experimental_enabled is True
-    assert (
-        is_feature_enabled("fleet", cfg.features, experimental_enabled=cfg.experimental_enabled)
-        is True
-    )
